added listOfLists class, added listview binding for files in directory
